requirements.txt
numpy
pandas
django-pandas
